pyproject.toml
[project]
name = "gh-space-shooter"
version = "0.1.0"
description = "A CLI tool that visualizes GitHub contribution graphs as gamified GIFs"
readme = "README.md"
authors = [
    { name = "zane", email = "czl970721@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "typer>=0.12.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "pillow>=10.0.0",
]

[project.scripts]
gh-space-shooter = "gh_space_shooter.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
